RAG Architecture/requirements.txt
faiss-cpu >= 1.11.0
sentence-transformers >= 4.1.0
langchain >= 0.3.25
pandas >= 2.2.3
numpy >= 2.1.2
torch >= 2.6.0
ollama >= 0.4.8